requirements.txt
# requirements.txt
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
sqlalchemy==2.0.19
openai==1.3.0
python-telegram-bot==20.3
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
email-validator==2.0.0
pytz==2023.3